// apps/desktop/src/routes/_dashboard/$user/-components/user/__generated__/UserInfo.graphql.ts
export type SocialAccountProvider =
  | "FACEBOOK"
  | "GENERIC"
  | "HOMETOWN"
  | "INSTAGRAM"
  | "LINKEDIN"
  | "MASTODON"
  | "NPM"
  | "REDDIT"
  | "TWITCH"
  | "TWITTER"
  | "YOUTUBE"
  | "%future added value";
import { FragmentRefs } from "relay-runtime";
export type UserInfo$data = {
  readonly avatarUrl: string;
  readonly bio: string | null | undefined;
  readonly company: string | null | undefined;
  readonly createdAt: string;
  readonly email: string;
  readonly id: string;
  readonly location: string | null | undefined;
  readonly login: string;
  readonly name: string | null | undefined;
  readonly socialAccounts: {
    readonly nodes:
      | ReadonlyArray<
          | {
              readonly displayName: string;
              readonly provider: SocialAccountProvider;
              readonly url: string;
            }
          | null
          | undefined
        >
      | null
      | undefined;
  };
  readonly topRepositories: {
    readonly nodes:
      | ReadonlyArray<
          | {
              readonly primaryLanguage:
                | {
                    readonly color: string | null | undefined;
                    readonly id: string;
                    readonly name: string;
                  }
                | null
                | undefined;
            }
          | null
          | undefined
        >
      | null
      | undefined;
  };
  readonly twitterUsername: string | null | undefined;
  readonly url: string;
  readonly websiteUrl: string | null | undefined;
  readonly " $fragmentSpreads": FragmentRefs<"FollowUserButton_user">;
  readonly " $fragmentType": "UserInfo";
};
export type UserInfo$key = {
  readonly " $data"?: UserInfo$data;
  readonly " $fragmentSpreads": FragmentRefs<"UserInfo">;
};